Operations and Built-in function.
Math() is the simplest and most frequently used operation. This is used for evaluation of mathematical expressions.

The various mathematical operations are-

addition +
substraction -
multiplication *
divide /
power **, and many more.
Though its not required to import math() because its an inbuilt module, the use of import has been shown so that other such modules can be put to use which are not available as an in built module in python.

Typecasting- This used to change data types. Examples are given in the code cell below. f=7.5 i = int(f)

type() returns the data type of the parameter passed.

help() - The python help function is used to display the documentation of modules, functions, classes, keywords etc.