Lists Operations and Strings manipulation.
**List Operations :-**
Various functions for list modification are as follows-
1. append()	Adds an element at the end of the list
2. clear()	Removes all the elements from the list
3. copy()	Returns a copy of the list
4. count()	Returns the number of elements with the specified value
5. extend()	Add the elements of a list (or any iterable), to the end of the current list
6. index()	Returns the index of the first element with the specified value
7. insert()	Adds an element at the specified position
8. pop()	Removes the element at the specified position
9. remove()	Removes the item with the specified value
10. reverse()	Reverses the order of the list
11. sort()	Sorts the list
For sorting in descending order, sort() and remove() can be used simultaneously.

The above functions are applicable even for Sets.
pop() for Sets is not useful because Sets dosent have unique index number for its elements.

**String Manipulations :-**
1. Indexing- Each element (alphabet, numbers, spaces) in a string holds an unique index number starting from [0] as the first element.

2. Negative Indexing- starts from the last letter of the string which is assigned an index of [-1]. This method can be used to separate out the last letter of a string for some operation as and when required.

Various coded examples are shown in the code cell below.
An element will therefore have *two indexes* i.e a positive and a negative index. Both the following indexes corresponds to the same element 'a'.

3. String Splicing (applicable to lists as well):-
This method is used to separate out a part of a list/string and then work on it.
Syntax stringname[start:stop:step]
start- index from where to start
stop- index till where string has to be sliced. This index element is not included in splicing.
step- no. of elements to be skipped (optional)

Various examples of string splicing has been shown in the code below for better understanding.

REVERSE of a string- can be done using string splicing.
Examples has been shown in the code below.
Just one syntax and a whole string/list can be easily reversed.