Bubble sort
The most commonly used sorting technique is the Bubblesort method. This is also the most easiest one and can be coded easily.

Theory-

A list of elements to be sorted is given as an input.

In bubblesorting, the operation begins from the first element with index [0].

Step 1- The first and the second element are compared. Depending upon the sorting (ascending or descending order), if the elements are already in proper order, no change is made. If the elements are not arranged then, both the elements are swapped and saved in the list.

Step 2- Now the operation moves to the next element. Here, the second and the third element are compared and if required, they are swapped depending upon the condition.

Step 3- Comparison between 2 successive elements continue in similar manner till the last element is reached. Finally we get the sorted list after the last step.

Run the code below in order to have a better understanding of each code lines and also to get more information about them.