Selection sort
In this method, the entered list is sub-divided into 2 parts. A sorted list and the unsorted list.

We work on the unsorted list and an element is appended to the sorted list during each cycle of iteration.

Theory-

Step 1- The first indexed element of the unsorted list is assumed to be minimum.

Step 2- We look at each element of the unsorted list and find the minimum element (in case of ascending order)

Step 3- This minimum element is swapped with the first indexed element so that the minimum number comes at the beginning.

Step 4- This process continues in each iteration and so the formation of the sorted list continues until all the elements of the unsorted list is handled.

Let's see how it works in action with a list that contains the following elements: [3, 5, 1, 2, 4].

We begin with the unsorted list:

3 5 1 2 4

The unsorted section has all the elements. We look through each item and determine that 1 is the smallest element. So, we swap 1 with 3:

1 5 3 2 4

Of the remaining unsorted elements, [5, 3, 2, 4], 2 is the lowest number. We now swap 2 with 5:

1 2 3 5 4

This process continues until the list is sorted:

1 2 3 5 4

1 2 3 4 5

1 2 3 4 5

Run the code below in order to have a better understanding of each code lines and also to get more information about them.