Insertion sort
In this sorting method, the unsorted list is taken as input and operation begins from the element at index position [1]. This value is passed to a loop and compared with the previous element. If condition is satisfied, the element at index [1], is shifted to its correct position. Now the next element at index [2] is passed to the loop. If conditions are satisfied, it is shifted to the left till the looping condition is true. This process continues till all the elements are shifted and inserted at their correct position depending upon ascending or descending order of sorting.

Theory-

Step 1- Iterate over the input elements by growing the sorted array at each iteration.

Step 2- Compare the current element with the largest value available in the sorted array.

Step 3- If the current element is greater, then it leaves the element in its place and moves on to the next element else it finds its correct position in the sorted array and moves it to that position in the array.

Step 4- This is achieved by shifting all the elements towards the right, which are larger than the current element, in the sorted array to one position ahead.

Working -

Entered list - 3 7 5 6 9 compare 7 with 3, since 7 > 3, no shifting

Now compare 5 with 7, since 5 < 7, move 5 to the left. We get, 3 5 7 6 9. Now compare 5 with 3, since 5 > 3, no shifting.

Now compare 6 with 7, since 6 < 7, move 6 to left. We get, 3 5 6 7 9. Now compare 5 with 6, since 5 < 6, no shifting.

Now compare 7 with 9, since 7 < 9, no shifting.

Final list - 3 5 6 7 9

Run the code below in order to have a better understanding of each code lines and also to get more information about them.