Binary searching
Searching methods, as the name says, is a tool to search for an element in an entered list and also to find its position in the list for further operations as required. Binary search is a method to do it.

Prerequisite - To implement Binary search, data on which we are working should be a sorted data in ascending or descending order. In case data entered is unsorted, we first sort it and then implement Binary search. This sorted data should be known as the final position of the element is wrt this sorted data.

Sorting- This can be done using any of the methods explained above. However, we have an inbuilt function < list_name >.sort() which sorts the data in ascending order in just one line. So we use this function here.

Theory-

In Binary search, the sorted data is taken, and no. of elements is known to us.

Step 1- Initialise a booelan variable 'found' as False. Start a while loop which runs till element is not found.

Step 2- Find the middle index value.

Step 3- Compare the search element value with middle value. Since, data is sorted, its easy to guess where our search element lies. If search value < middle value, it lies in the first half. Otherwise it lies in the secoond half. The best case scenario is when search value lies in the middle.

Now each element in the respective half can be checked whether they match the search element value.

Run the code below in order to have a better understanding of each code lines and also to get more information about them.