Linear searching
Another searching method is the Linear Search method. Here, we do not require a pre-sorted list like we used in Binary Search.

The working process is pretty simple and direct in Linear search method.

Theory -

Step 1- We have a list (may be user input (or) randomly generated)

Step 2- Enter the search element from the user

Step 3- Check each element of the list and see if it matches the search element

Step 4- print the result

Run the code below in order to have a better understanding of each code lines and also to get more information about them.