numpy and random module
Using numpy and random modules

numpy.random.randint() is one of the function for doing random sampling in numpy. It returns an array of specified shape and fills it with random integers from low (inclusive) to high (exclusive), i.e. in the interval [low, high).

When to use :-

When we want the computer to pick a random number in a given range.

If we wanted a random integer, we can use the randint function

This was just a basic introduction which was required as we will use this in the next topic. There are many other methods available for different random list generation in the random module.

Run the code below in order to have a better understanding of each code lines and also to get more information about them.